+++ src/v2/services/implementations/XcmEvmService.ts
@@ -8,10 +8,10 @@ import { getPubkeyFromSS58Addr } from 'src/hooks/helper/addressUtils';
 import { isValidAddressPolkadotAddress } from 'src/hooks/helper/plasmUtils';
 import { getEvmProvider } from 'src/hooks/helper/wallet';
 import { getEvmGas } from 'src/modules/gas-api';
-import { ethWalletChains, relaychainParaId, xcmChainObj } from 'src/modules/xcm';
-import { Chain } from 'src/v2/models/XcmModels';
+import { relaychainParaId, xcmChainObj } from 'src/modules/xcm';
 import { Guard } from 'src/v2/common';
 import { BusyMessage, ExtrinsicStatusMessage, IEventAggregator } from 'src/v2/messaging';
+import { Chain, ethWalletChains } from 'src/v2/models';
 import { IGasPriceProvider, IXcmEvmService, TransferParam } from 'src/v2/services';
 import { Symbols } from 'src/v2/symbols';
 import Web3 from 'web3';
